Add cd to hugo server start function
If you are in ~ and start hugo server, even with a --source and config
file specified, it gets confused, and does not work 100% correctly.

No error is seen, but it is evident in how it serves the content, incorr-

Added cd /Users/rcogley/Repo.../eSolia before starting hugo server, and
this seems to take care of it.